Added a login page that requests the /login API.
This doesn't work well yet. The request is made but then there are
some problems. First problem is the cookies aren't kept and resend
with future requests which renders this completely useless. The
second problem is more of a TODO, we must check all the requests to
the API and if they 401 Unauthorized we should redirect the user to
the login page.